requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.3
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
